refactor: LlmMatchScoreService 내부 result 타입 패키지 분리 (#144)
