+++ src/main/java/com/kusitms/kkium/jd/utils/llm/result/HighlightKeyword.java
@@ -1,0 +1,5 @@
+package com.kusitms.kkium.jd.utils.llm.result;
+
+import java.util.List;
+
+public record HighlightKeyword(String keyword, List<String> sources) {}
