+++ src/main/java/com/kusitms/kkium/jd/utils/llm/result/LlmMatchResult.java
@@ -1,0 +1,5 @@
+package com.kusitms.kkium.jd.utils.llm.result;
+
+import java.util.Map;
+
+public record LlmMatchResult(Map<Long, Integer> usageScores, int applicationScore) {}
