+++ src/main/java/com/kusitms/kkium/jd/utils/llm/result/LlmQuestionMatchResult.java
@@ -1,0 +1,5 @@
+package com.kusitms.kkium.jd.utils.llm.result;
+
+import java.util.Map;
+
+public record LlmQuestionMatchResult(Map<Long, Integer> usageScores) {}
